requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
docopt==0.6.2
idna==2.10
num2words==0.5.10
python-obscene-words-filter==0.1.6
pytz==2021.1
requests==2.25.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
vk-api==11.9.1
